+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Add validate function to client and server to confirm that new results are consistent with past submissions
+- Add experimental support for GPU acceleration with CUDA
 
 ## Nice v3.2.9